+++ services/storage/java/androidx/test/services/storage/provider/InternalUseOnlyFilesContentProvider.java
@@ -15,39 +15,23 @@
  */
 package androidx.test.services.storage.provider;
 
-import android.os.Environment;
-import android.util.Log;
+import android.content.Context;
 import androidx.test.services.storage.TestStorageConstants;
 import androidx.test.services.storage.file.HostedFile;
 import java.io.File;
 
 /** Hosts an SD Card directory for the test framework to read/write internal files to. */
 public final class InternalUseOnlyFilesContentProvider extends AbstractFileContentProvider {
-  private static final String TAG = "InternalUseOnlyFilesContentProvider";
 
-  private final File outputDirectory;
-
-  public InternalUseOnlyFilesContentProvider() {
-    super(
-        new File(HostedFile.getRootDirectory(), TestStorageConstants.ON_DEVICE_PATH_INTERNAL_USE),
-        AbstractFileContentProvider.Access.READ_WRITE);
-    outputDirectory =
-        new File(HostedFile.getRootDirectory(), TestStorageConstants.ON_DEVICE_PATH_INTERNAL_USE);
+  @Override
+  protected File getHostedDirectory(Context context) {
+    return new File(
+        HostedFile.getOutputRootDirectory(context),
+        TestStorageConstants.ON_DEVICE_PATH_INTERNAL_USE);
   }
 
   @Override
-  protected boolean onCreateHook() {
-    if (!Environment.getExternalStorageState().equals(Environment.MEDIA_MOUNTED)) {
-      Log.e(TAG, "sdcard in bad state: " + Environment.getExternalStorageState());
-      return false;
-    } else {
-      if (!outputDirectory.exists()) {
-        if (!outputDirectory.mkdirs()) {
-          Log.e(TAG, String.format("'%s': could not create output dir! ", outputDirectory));
-          return false;
-        }
-      }
-      return true;
-    }
+  protected Access getAccess() {
+    return AbstractFileContentProvider.Access.READ_WRITE;
   }
 }