+++ services/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 **Bug Fixes**
 
+* TestStorage: use local cache dir to store output files when running as non system user
+
 **New Features**
 
 **Breaking Changes**